backend/requirements.txt
django-extensions
djangorestframework
django-cors-headers
pyotp
geopy
qrcode
Pillow
djangorestframework-simplejwt
paypalrestsdk